Managed to partially fix the raycast script.
It now does a 1 way sonar raycast by which it left to right but not right to left fully yet.
It also itsn't bound to the player's position so I still need to fix that as of yet.
but its progress nevertheless.

tomorow I should hope to completely fix this bug and hopefully get the flashlight scripts all working as they should and then will work on AI.